services/Gateway/requirements.txt
fastapi==0.115.5
python-dotenv==1.0.1
uvicorn==0.32.0
httpx==0.27.2
